requirements.txt
azureml.core==1.41.0
azureml.pipeline==1.41.0
azureml.dataprep.native==38.0.0
argparse==1.4.0
numpy==1.22.3
pandas==1.4.2
